moved related widget in pages
